シェーダーの組み合わせを登録、取得できるようにした
[概要]
・シェーダーの組み合わせを保持する[ShaderProgram]を作成して、[ShaderManager]で登録、取得できるようにした
・設定されているシェーダーのバインドを一斉に行うメソッドを持っている
・デフォルト設定のプログラムを取得できるようにした
・上記変更の実行確認として[TestRenderer]でShaderProgramを取得してシェーダーを設定するようにした

+++ DX11Framework-2025/Code/Source/Framework/Shaders/VertexShader.cpp
@@ -62,7 +62,7 @@ bool VertexShader::CreateShader(ID3D11Device& _device, const ShaderInfo _shaderI
 	// 入力レイアウトの作成
 	hr = _device.CreateInputLayout(
 		ShaderCommon::LayoutDescs[static_cast<size_t>(_shaderInfo.layoutType)].data(),
-		ShaderCommon::LayoutDescs[static_cast<UINT>(_shaderInfo.layoutType)].size(),
+		static_cast<UINT>(ShaderCommon::LayoutDescs[static_cast<size_t>(_shaderInfo.layoutType)].size()),
 		this->blob->GetBufferPointer(),
 		this->blob->GetBufferSize(),
 		this->inputLayout.GetAddressOf()
